pyproject.toml
[project]
name = "hgse-event-calendar"
version = "0.1.0"
description = "Weekly digest of Helsinki GSE seminar events"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "supabase>=2.0",
    "python-dotenv>=1.0",
    "sendgrid>=6.0",
    "flask>=3.0",
]

[project.scripts]
scrape-events = "scraper.cli:main"
